+++ cbot_actuators/launch/cbot_actuators.launch
@@ -1,0 +1,6 @@
+<?xml version="1.0"?>
+<launch>
+    <rosparam file="$(find cbot_actuators)/config/params.yaml" command="load" />
+    <node name="thruster_node" pkg="cbot_actuators" type="thruster_node" output="screen">
+    </node>
+</launch>
